+++ firmware/src/config/default/initialization.c
@@ -139,6 +139,9 @@ void SYS_Initialize ( void* data )
 
     EVSYS_Initialize();
 
+    SERCOM5_USART_Initialize();
+
+    ADC_Initialize();
 	BSP_Initialize();