+++ firmware/src/config/default/peripheral/clock/plib_clock.c
@@ -95,7 +95,14 @@ void CLOCK_Initialize (void)
     GCLK0_Initialize();
 
 
+    /* Selection of the Generator and write Lock for SERCOM5_CORE */
+    GCLK_REGS->GCLK_CLKCTRL = GCLK_CLKCTRL_ID(25) | GCLK_CLKCTRL_GEN(0x0)  | GCLK_CLKCTRL_CLKEN_Msk;
+    /* Selection of the Generator and write Lock for ADC */
+    GCLK_REGS->GCLK_CLKCTRL = GCLK_CLKCTRL_ID(30) | GCLK_CLKCTRL_GEN(0x0)  | GCLK_CLKCTRL_CLKEN_Msk;
     
+    /* Configure the APBC Bridge Clocks */
+    PM_REGS->PM_APBCMASK = 0x10080;
+
 
     /*Disable RC oscillator*/
     SYSCTRL_REGS->SYSCTRL_OSC8M = 0x0;