+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All major changes to the project are recorded here
 
+## 2020-07-26 : Added code for ADC and relay control.
+
+[Added]
+
+- Code for ADC
+- Block for ADC
+
 ## 2020-07-22 : Initial Commit
 
 ## [Added]